GeoJSON - refactor feature/geometry parsing
